todo-app done

+++ src/app/app.component.ts
@@ -1,6 +1,8 @@
-import { Component, HostBinding} from '@angular/core';
+import { Component, HostBinding, inject } from '@angular/core';
 import { Todo } from './Model/Todo';
 import { CdkDragDrop, moveItemInArray } from '@angular/cdk/drag-drop';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogBoxComponent } from './dialog-box/dialog-box.component';
 
 @Component({
   selector: 'app-root',
@@ -13,10 +15,9 @@ export class AppComponent {
   currActiveFilterType: string = 'all';
   isDarkMode!: boolean;
   dragging = false;
-  screenWidth!: number;
+  constructor(private dialog: MatDialog) {}
 
   ngOnInit() {
-    this.screenWidth = window.innerWidth;
     const storedTodos = localStorage.getItem('todos');
     this.todos = storedTodos
       ? (JSON.parse(storedTodos) as Todo[])
@@ -38,7 +39,6 @@ export class AppComponent {
     return this.isDarkMode;
   }
 
-
   getItemsLeft() {
     return this.todos.filter((todo) => todo.isActive).length;
   }
@@ -46,18 +46,45 @@ export class AppComponent {
     todo.isActive = !todo.isActive;
     localStorage.setItem('todos', JSON.stringify(this.todos));
   }
-  removeTodo(index: number) {
-    this.todos.splice(index, 1);
-    this.copyOfTodos.splice(index, 1);
-    localStorage.setItem('todos', JSON.stringify(this.todos));
+  removeTodo(isRemoving: boolean, index: number) {
+    if (isRemoving) {
+      const dialogRef = this.dialog.open(DialogBoxComponent, {
+        maxWidth: '300px',
+        data: {
+          title: 'Confirm Delete',
+          message: 'Are you sure you want to delete this todo?',
+          isDelete: true,
+        },
+      });
+
+      dialogRef.afterClosed().subscribe({
+        next: (result: boolean) => {
+          if (result) {
+            this.todos.splice(index, 1);
+            this.copyOfTodos.splice(index, 1);
+            localStorage.setItem('todos', JSON.stringify(this.todos));
+          }
+        },
+      });
+    }
   }
   addTodo(inputTodo: string, input: HTMLInputElement) {
     if (inputTodo) {
       const result = this.todos.find(
         (todo) => todo.todo.toLowerCase() === inputTodo.toLowerCase()
       );
-      if (result) console.log('Todo Already existed');
-      else {
+      if (result) {
+        this.dialog.open(DialogBoxComponent, {
+          maxWidth: '300px',
+          data: {
+            title: 'Todo Already Exists',
+            message:
+              'The todo you entered already exists. Please enter a different todo.',
+            isDelete: false,
+          },
+        });
+        return;
+      } else {
         this.todos.unshift({ todo: inputTodo, isActive: true });
         input.value = '';
       }
